storage-node: introduce CLI typescript project
