+++ storage-node/package.json
@@ -32,7 +32,8 @@
   ],
   "scripts": {
     "test": "wsrun --serial test",
-    "lint": "eslint --ignore-path .gitignore ."
+    "lint": "eslint --ignore-path .gitignore .",
+    "postinstall": "tsc --build"
   },
   "devDependencies": {
     "eslint": "^5.16.0",
@@ -41,6 +42,9 @@
     "eslint-plugin-babel": "^5.3.1",
     "eslint-plugin-prettier": "^3.1.4",
     "prettier": "^2.0.5",
-    "wsrun": "^3.6.5"
+    "wsrun": "^3.6.5",
+    "typescript": "^3.9.6",
+    "@types/chai": "^4.2.11",
+    "@types/mocha": "^7.0.2"
   }
 }
